// src/payload-generated-schema.ts
export const enum_users_role = pgEnum("enum_users_role", [
  "admin",
  "editor",
  "author",
  "viewer",
]);
export const enum_posts_status = pgEnum("enum_posts_status", [
  "draft",
  "review",
  "scheduled",
  "published",
  "archived",
]);
// ...
export const users = pgTable(
  "users",
  {
    id: serial("id").primaryKey(),
    name: varchar("name").notNull(),
    role: enum_users_role("role").notNull().default("author"),
    avatar: integer("avatar_id").references(() => media.id, {
      onDelete: "set null",
    }),
    isActive: boolean("is_active").default(true),
    lastLoginAt: timestamp("last_login_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    }),
    updatedAt: timestamp("updated_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp("created_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    email: varchar("email").notNull(),
    resetPasswordToken: varchar("reset_password_token"),
    resetPasswordExpiration: timestamp("reset_password_expiration", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    }),
    salt: varchar("salt"),
    hash: varchar("hash"),
    loginAttempts: numeric("login_attempts", { mode: "number" }).default(0),
    lockUntil: timestamp("lock_until", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    }),
  },
  (columns) => [
    index("users_avatar_idx").on(columns.avatar),
    index("users_updated_at_idx").on(columns.updatedAt),
    index("users_created_at_idx").on(columns.createdAt),
    uniqueIndex("users_email_idx").on(columns.email),
  ],
);

export const authors = pgTable(
  "authors",
  {
    id: serial("id").primaryKey(),
    name: varchar("name").notNull(),
    slug: varchar("slug").notNull(),
    bio: varchar("bio"),
    avatar: integer("avatar_id").references(() => media.id, {
      onDelete: "set null",
    }),
    email: varchar("email"),
    socialLinks_facebook: varchar("social_links_facebook"),
    socialLinks_twitter: varchar("social_links_twitter"),
    socialLinks_instagram: varchar("social_links_instagram"),
    user: integer("user_id").references(() => users.id, {
      onDelete: "set null",
    }),
    isActive: boolean("is_active").default(true),
    updatedAt: timestamp("updated_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp("created_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
  },
  (columns) => [
    uniqueIndex("authors_slug_idx").on(columns.slug),
    index("authors_avatar_idx").on(columns.avatar),
    index("authors_user_idx").on(columns.user),
    index("authors_updated_at_idx").on(columns.updatedAt),
    index("authors_created_at_idx").on(columns.createdAt),
  ],
);

export const posts = pgTable(
  "posts",
  {
    id: serial("id").primaryKey(),
    title: varchar("title"),
    slug: varchar("slug"),
    excerpt: varchar("excerpt"),
    content: jsonb("content"),
    featuredImage: integer("featured_image_id").references(() => media.id, {
      onDelete: "set null",
    }),
    category: integer("category_id").references(() => categories.id, {
      onDelete: "set null",
    }),
    author: integer("author_id").references(() => authors.id, {
      onDelete: "set null",
    }),
    status: enum_posts_status("status").default("draft"),
    publishedAt: timestamp("published_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    }),
    scheduledAt: timestamp("scheduled_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    }),
    seoTitle: varchar("seo_title"),
    seoDescription: varchar("seo_description"),
    canonicalUrl: varchar("canonical_url"),
    ogImage: integer("og_image_id").references(() => media.id, {
      onDelete: "set null",
    }),
    isFeatured: boolean("is_featured").default(false),
    isBreakingNews: boolean("is_breaking_news").default(false),
    allowIndex: boolean("allow_index").default(true),
    sourceName: varchar("source_name"),
    sourceUrl: varchar("source_url"),
    readingTime: numeric("reading_time", { mode: "number" }),
    views: numeric("views", { mode: "number" }).default(0),
    updatedAt: timestamp("updated_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp("created_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    _status: enum_posts_status("_status").default("draft"),
  },
  (columns) => [
    uniqueIndex("posts_slug_idx").on(columns.slug),
    index("posts_featured_image_idx").on(columns.featuredImage),
    index("posts_category_idx").on(columns.category),
    index("posts_author_idx").on(columns.author),
    index("posts_og_image_idx").on(columns.ogImage),
    index("posts_updated_at_idx").on(columns.updatedAt),
    index("posts_created_at_idx").on(columns.createdAt),
    index("posts__status_idx").on(columns._status),
  ],
);
// ...
export const categories = pgTable(
  "categories",
  {
    id: serial("id").primaryKey(),
    name: varchar("name").notNull(),
    slug: varchar("slug").notNull(),
    description: varchar("description"),
    parent: integer("parent_id").references((): AnyPgColumn => categories.id, {
      onDelete: "set null",
    }),
    seoTitle: varchar("seo_title"),
    seoDescription: varchar("seo_description"),
    order: numeric("order", { mode: "number" }).default(0),
    isActive: boolean("is_active").default(true),
    updatedAt: timestamp("updated_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp("created_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
  },
  (columns) => [
    uniqueIndex("categories_slug_idx").on(columns.slug),
    index("categories_parent_idx").on(columns.parent),
    index("categories_updated_at_idx").on(columns.updatedAt),
    index("categories_created_at_idx").on(columns.createdAt),
  ],
);
// ...
export const media = pgTable(
  "media",
  {
    id: serial("id").primaryKey(),
    alt: varchar("alt").notNull(),
    caption: varchar("caption"),
    credit: varchar("credit"),
    folder: varchar("folder").default("/"),
    updatedAt: timestamp("updated_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp("created_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    url: varchar("url"),
    thumbnailURL: varchar("thumbnail_u_r_l"),
    filename: varchar("filename"),
    mimeType: varchar("mime_type"),
    filesize: numeric("filesize", { mode: "number" }),
    width: numeric("width", { mode: "number" }),
    height: numeric("height", { mode: "number" }),
    focalX: numeric("focal_x", { mode: "number" }),
    focalY: numeric("focal_y", { mode: "number" }),
    sizes_thumbnail_url: varchar("sizes_thumbnail_url"),
    sizes_thumbnail_width: numeric("sizes_thumbnail_width", { mode: "number" }),
    sizes_thumbnail_height: numeric("sizes_thumbnail_height", {
      mode: "number",
    }),
    sizes_thumbnail_mimeType: varchar("sizes_thumbnail_mime_type"),
    sizes_thumbnail_filesize: numeric("sizes_thumbnail_filesize", {
      mode: "number",
    }),
    sizes_thumbnail_filename: varchar("sizes_thumbnail_filename"),
    sizes_card_url: varchar("sizes_card_url"),
    sizes_card_width: numeric("sizes_card_width", { mode: "number" }),
    sizes_card_height: numeric("sizes_card_height", { mode: "number" }),
    sizes_card_mimeType: varchar("sizes_card_mime_type"),
    sizes_card_filesize: numeric("sizes_card_filesize", { mode: "number" }),
    sizes_card_filename: varchar("sizes_card_filename"),
    sizes_hero_url: varchar("sizes_hero_url"),
    sizes_hero_width: numeric("sizes_hero_width", { mode: "number" }),
    sizes_hero_height: numeric("sizes_hero_height", { mode: "number" }),
    sizes_hero_mimeType: varchar("sizes_hero_mime_type"),
    sizes_hero_filesize: numeric("sizes_hero_filesize", { mode: "number" }),
    sizes_hero_filename: varchar("sizes_hero_filename"),
    sizes_og_url: varchar("sizes_og_url"),
    sizes_og_width: numeric("sizes_og_width", { mode: "number" }),
    sizes_og_height: numeric("sizes_og_height", { mode: "number" }),
    sizes_og_mimeType: varchar("sizes_og_mime_type"),
    sizes_og_filesize: numeric("sizes_og_filesize", { mode: "number" }),
    sizes_og_filename: varchar("sizes_og_filename"),
  },
  (columns) => [
    index("media_updated_at_idx").on(columns.updatedAt),
    index("media_created_at_idx").on(columns.createdAt),
    uniqueIndex("media_filename_idx").on(columns.filename),
    index("media_sizes_thumbnail_sizes_thumbnail_filename_idx").on(
      columns.sizes_thumbnail_filename,
    ),
    index("media_sizes_card_sizes_card_filename_idx").on(
      columns.sizes_card_filename,
    ),
    index("media_sizes_hero_sizes_hero_filename_idx").on(
      columns.sizes_hero_filename,
    ),
    index("media_sizes_og_sizes_og_filename_idx").on(columns.sizes_og_filename),
  ],
);